+++ libs/base_std/src/code/result/mod.rs
@@ -1,0 +1,7 @@
+// devela_base_std::code::result
+//
+#![doc = crate::_DOC_CODE_RESULT!()]
+//
+
+#[cfg(test)]
+mod unwrap_test; // tests for unwrap!