+++ src/code/result/opt_res/mod.rs
@@ -8,16 +8,12 @@
 mod ext_option; // ExtOption
 mod ext_result; // ExtResult
 mod fmt; // OptionFmt, OptionFmtOr, OptionFmtOrElse
-mod opt_res; // serr, sok, OptRes
-mod unwrap; // unwrap!
 
 crate::structural_mods! { // _mods, _always
     _mods {
-        pub use super::{
-            ext_option::*, ext_result::*, fmt::*, opt_res::*, unwrap::*,
-        };
+        pub use super::{ext_option::*, ext_result::*, fmt::*};
     }
     _always {
-        pub use super::{ext_option::*, ext_result::*, fmt::*, opt_res::*, unwrap::*};
+        pub use super::{ext_option::*, ext_result::*, fmt::*};
     }
 }